+++ src/models/index.ts
@@ -1,3 +1,5 @@
-import {JwtPayload, SCOPE_DIVIDER} from './JwtPayload'
-
-export {JwtPayload, SCOPE_DIVIDER}
+export {JwtPayload} from './JwtPayload'
+export {SSOParams} from './SSOParams'
+export {KeyValueI} from '../models/KeyValueI'
+export {SSOParamsI} from '../models/SSOParamsI'
+export {JwtPayloadI} from '../models/JwtPayloadI'